door&light unification, utils, moved home creation into HomeBuilder
