+++ src/main/java/ru/sbt/mipt/oop/devices/Light.java
@@ -1,16 +1,12 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.devices;
 
 public class Light {
-    private boolean isOn;
     private final String id;
+    private boolean isOn;
 
     public Light(String id, boolean isOn) {
         this.id = id;
         this.isOn = isOn;
-    }
-
-    public boolean isOn() {
-        return isOn;
     }
 
     public String getId() {
@@ -20,4 +16,5 @@ public class Light {
     public void setOn(boolean on) {
         isOn = on;
     }
+
 }
